+++ src/features/boolean-algebra/components/InteractiveCalculator.jsx
