+++ src/features/kmap/components/InputControls.jsx
