+++ src/features/kmap/components/SimplifiedExpression.jsx
