+++ src/features/kmap/components/TruthTableDisplay.jsx
