+++ src/features/kmap/components/WhiteboardAnimation.jsx
